Cleaned App.xaml.cs | Organized other stuff | Removed useless calls. | Fixed Build for Github | Some temp icons are included will be removed later | WKitGlobal introduced : Holds helpers for global stuff. And A AppHelper to register the ViewModels there instead. | Added Homepage button on ribbon | Removed old buttons on titlebar | Added Title |
