+++ WolvenKit/Views/Editors/CodeEditor/CodeEditorView.xaml
@@ -22,7 +22,7 @@
         </ResourceDictionary>
     </catel:DataWindow.Resources>
 
-    <Border hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" Width="auto" Height="auto" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
+    <Border hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" Width="auto" Height="auto" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
 
         <Grid Background="{DynamicResource Fluent.Ribbon.Brushes.Backstage.Background}">
