+++ WolvenKit/Views/HomePage/Pages/WelcomePageView.xaml
@@ -110,7 +110,7 @@
     </catel:UserControl.Resources>
 
    
-        <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}"  MinWidth="300" MinHeight="300" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="5,5,2,5">
+        <Grid hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}"  MinWidth="300" MinHeight="300" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="5,5,2,5">
 
 
         <Grid.RowDefinitions>
@@ -221,7 +221,7 @@
         <Grid Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="2" >
             <hc:TransitioningContentControl TransitionMode="Right2LeftWithFade">
 
-                <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}"  Background="{DynamicResource MahApps.Brushes.ThemeBackground}"  >
+                <Grid  hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}"  Background="{DynamicResource MahApps.Brushes.ThemeBackground}"  >
                 <Grid Margin="3,0,3,0">
                         <StackPanel>
                             
@@ -233,7 +233,7 @@
                         </StackPanel>
                         <hc:Divider Margin="32,0,32,0"></hc:Divider>
                     <hc:TransitioningContentControl MinHeight="250" MaxHeight="200" Margin="0,0,0,0">
-                            <hc:CoverView  ShowContent="False" BorderBrush="{x:Null}"  Grid.RowSpan="1" ItemContentHeight="55" Background="{DynamicResource MahApps.Brushes.ThemeBackground}" Margin="15" BorderThickness="1,1,1,1" ItemMargin="15,8,15,8" ItemHeight="200"  ItemWidth="350">
+                                <hc:CoverView                        ItemsSource="{Binding RecentlyUsedItems}"  ShowContent="False" BorderBrush="{x:Null}"  Grid.RowSpan="1" ItemContentHeight="55" Background="{DynamicResource MahApps.Brushes.ThemeBackground}" Margin="15" BorderThickness="1,1,1,1" ItemMargin="15,8,15,8" ItemHeight="200"  ItemWidth="350">
 
                                 <hc:CoverView.ItemHeaderTemplate>
                                     <DataTemplate>
@@ -242,7 +242,7 @@
                                                 <Border Background="{DynamicResource Fluent.Ribbon.Brushes.BackstageTabControl.ItemsPanelBackground}" CornerRadius="15,2,2,2">
                                                     <hc:Tag  Height="40" HorizontalContentAlignment="Left" HorizontalAlignment="Left" Background="{x:Null}"  MinWidth="350" MaxWidth="350" Margin="3" ShowCloseButton="False">
                                                         <hc:Tag.Content>
-                                                            <TextBlock HorizontalAlignment="Left" Margin="0,0,0,0" Foreground="{DynamicResource MahApps.Brushes.ThemeForeground}" FontSize="16"   Text="Project Name"></TextBlock>
+                                                            <TextBlock HorizontalAlignment="Left" Margin="0,0,0,0" Foreground="{DynamicResource MahApps.Brushes.ThemeForeground}" FontSize="16"   Text="{Binding Name}"></TextBlock>
                                                         </hc:Tag.Content>
                                                         <hc:Tag.Header>
                                                             <Border>
@@ -271,15 +271,23 @@
                                                             <hc:Divider Margin="5,15" LineStroke="{DynamicResource MahApps.Brushes.Accent3}" LineStrokeThickness="2" Width="15"/>
 
                                                             <Border VerticalAlignment="Center" Height="25" Background="{DynamicResource Fluent.Ribbon.Brushes.BackstageTabControl.ItemsPanelBackground}" CornerRadius="5,2,2,2">
-                                                                <TextBlock Width="280" Margin="5" VerticalAlignment="Center"  Text="Project Type : "/>
+                                                                    <StackPanel Orientation="Horizontal">
+                                                                        <TextBlock Width="74" Margin="5" VerticalAlignment="Center"  Text="Project Type : "/>
+                                                                        <TextBlock Width="280" Margin="5" VerticalAlignment="Center"  Text="NOTHING TO BIND"/>
+
+                                                                    </StackPanel>
                                                             </Border>
 
                                                         </StackPanel>
                                                         <StackPanel Margin="5" Orientation="Horizontal">
                                                             <hc:Divider Margin="5,15" LineStroke="{DynamicResource MahApps.Brushes.Accent3}" LineStrokeThickness="2" Width="15"/>
                                                             <Border VerticalAlignment="Center" Height="25" Background="{DynamicResource Fluent.Ribbon.Brushes.BackstageTabControl.ItemsPanelBackground}" CornerRadius="5,2,2,2">
-                                                                <TextBlock Width="280" Margin="5" VerticalAlignment="Center"  Text="Last Modified : "/>
-                                                            </Border>
+                                                                    <StackPanel Orientation="Horizontal">
+                                                                        <TextBlock Width="80" Margin="5" VerticalAlignment="Center"  Text="Last Modified : "/>
+                                                                        <TextBlock Width="280" Margin="5" VerticalAlignment="Center"  Text="{Binding DateTime}"/>
+
+                                                                    </StackPanel>
+                                                                </Border>
                                                         </StackPanel>
 
                                                     </StackPanel>
@@ -309,15 +317,7 @@
                                     </DataTemplate>
 
                                 </hc:CoverView.ItemHeaderTemplate>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
-                                <hc:CoverViewItem ></hc:CoverViewItem>
+                          
                             </hc:CoverView>
                             </hc:TransitioningContentControl>
                         </StackPanel>
