+++ WolvenKit/Views/HomePage/Pages/WikiPageView.xaml
@@ -6,7 +6,7 @@
                    xmlns:hc="https://handyorg.github.io/handycontrol"
                    IsVisibleChanged="UserControl_IsVisibleChanged">
 
-    <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}"  Background="{DynamicResource MahApps.Brushes.ThemeBackground}" MinWidth="300" MinHeight="300" >
+    <Grid hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}"  Background="{DynamicResource MahApps.Brushes.ThemeBackground}" MinWidth="300" MinHeight="300" >
         <Grid Margin="3,75,3,0">
             <hc:TransitioningContentControl TransitionMode="Right2LeftWithFade">
