+++ WolvenKit/Views/SettingsPages/GeneralSettingsView.xaml
@@ -7,7 +7,7 @@
                    IsVisibleChanged="UserControl_IsVisibleChanged"
                    >
 
-    <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" Background="{DynamicResource Fluent.Ribbon.Brushes.Backstage.Background}">
+    <Grid hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" Background="{DynamicResource Fluent.Ribbon.Brushes.Backstage.Background}">
         <hc:SideMenu Width="150" SideBrush="{DynamicResource MahApps.Brushes.Accent}" HorizontalAlignment="Left" >
 
             <hc:SideMenuItem x:Name="GlobalSubItem" Selected="GlobalSubItem_Selected" >
