+++ WolvenKit/Views/SettingsPages/SubPages/General/LoggingSubSettingsView.xaml
@@ -5,7 +5,7 @@
                    xmlns:hc="https://handyorg.github.io/handycontrol"
                    xmlns:catel="http://schemas.catelproject.com">
 
-    <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}">
+    <Grid hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}">
 
 
         <Grid.ColumnDefinitions>
