+++ WolvenKit/Views/Tools/AssetBrowser/AssetBrowserView.xaml
@@ -16,7 +16,7 @@
         </ResourceDictionary>
 
     </catel:DataWindow.Resources>
-    <Border hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
+    <Border hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
 
         <Grid>
             <Grid.RowDefinitions>
