+++ WolvenKit/Views/Tools/AudioTool/AudioToolView.xaml
@@ -148,7 +148,7 @@
 
     </catel:DataWindow.Resources>
 
-    <Border hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" Width="auto" Height="auto" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
+    <Border hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" Width="auto" Height="auto" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
 
         <Grid Background="{DynamicResource Fluent.Ribbon.Brushes.Backstage.Background}">
