+++ WolvenKit/Views/UserControlHostWindowView.xaml
@@ -23,7 +23,7 @@
 
     </catel:DataWindow.Resources>
     <Grid>
-        <Border hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" MinWidth="600" MinHeight="600" MaxWidth="1200" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
+        <Border hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" MinWidth="600" MinHeight="600" MaxWidth="1200" BorderBrush="{DynamicResource MahApps.Brushes.Accent4}" BorderThickness="1">
 
             <Grid Background="{DynamicResource Fluent.Ribbon.Brushes.Backstage.Background}">
