+++ WolvenKit/Views/Wizards/FeedbackWizardView.xaml
@@ -11,7 +11,7 @@
         </ResourceDictionary>
 
     </catel:UserControl.Resources>
-    <Grid hc:ThemeManager.AccentColor="{DynamicResource MahApps.Brushes.Accent3}" MinWidth="300"  HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="5,5,2,5">
+    <Grid hc:ThemeManager.RequestedAccentColor="{DynamicResource MahApps.Brushes.Accent3}" MinWidth="300"  HorizontalAlignment="Stretch" VerticalAlignment="Center" Margin="5,5,2,5">
 
 
         <Grid.RowDefinitions>
