+++ WolvenKit/WolvenKit.csproj
@@ -36,9 +36,14 @@
   <ItemGroup>
     <None Remove="Resources\Icons\3.png" />
     <None Remove="Resources\Icons\Homelogo.png" />
+    <None Remove="Resources\Icons\SeeTroughIcon.png" />
     <None Remove="Resources\Icons\Somthlikedis.png" />
+    <None Remove="Resources\Icons\TinyTest.png" />
     <None Remove="Resources\Icons\whitelogo.png" />
     <None Remove="Wkit_Icon_Flattened_256x.ico" />
+  </ItemGroup>
+  <ItemGroup>
+    <Resource Include="Resources\Icons\SeeTroughIcon.png" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\Icons\Somthlikedis.png" />
@@ -67,7 +72,7 @@
     <PackageReference Include="CsvHelper" Version="16.1.0" />
     <PackageReference Include="Dirkster.MLib" Version="1.3.1.2" />
     <PackageReference Include="FastMember" Version="1.5.0" />
-    <PackageReference Include="HandyControls" Version="3.1.0-rc.1" />
+    <PackageReference Include="HandyControls" Version="3.1.1" />
     <PackageReference Include="MahApps.Metro" Version="2.4.3" />
     <PackageReference Include="MahApps.Metro.IconPacks.BootstrapIcons" Version="4.8.0" />
     <PackageReference Include="MahApps.Metro.IconPacks.BoxIcons" Version="4.8.0" />
@@ -224,6 +229,7 @@
     <Resource Include="Resources\Icons\3.png" />
     <Resource Include="Resources\Icons\Homelogo.png" />
     <Resource Include="Resources\Icons\Logo.ico" />
+    <Resource Include="Resources\Icons\TinyTest.png" />
     <Resource Include="Resources\Icons\whitelogo.png" />
     <Resource Include="Resources\Images\Files\apb_32x.png" />
     <Resource Include="Resources\Images\Files\BlankFile_16x.png" />
@@ -340,12 +346,5 @@
     <None Update="testmp3.mp3">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="Properties\DesignTimeResources.xaml" Condition="'$(DesignTime)'=='true' OR ('$(SolutionPath)'!='' AND Exists('$(SolutionPath)') AND '$(BuildingInsideVisualStudio)'!='true' AND '$(BuildingInsideExpressionBlend)'!='true')">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-      <ContainsDesignTimeResources>true</ContainsDesignTimeResources>
-    </Page>
   </ItemGroup>
 </Project>
